services/umls-mcp-server/pyproject.toml
[project]
name = "umls-mcp-server"
version = "0.1.0"
description = "UMLS MCP server exposing concept search, linking, and semantic type tools via FastMCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diskcache>=5.6.3",
    "fastmcp",
    "httpx",
    "platformdirs>=4.5.1",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/umls_mcp_server"]

[tool.ruff]
lint.select = ["E", "F", "W", "C", "N", "I", "D"]
lint.ignore = ["E203", "D203", "D213", "D100", "D413", "D104"]
lint.pydocstyle.convention = "google"
target-version = "py312"
line-length = 88
exclude = ["venv", ".venv", "docs"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D102", "D103"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "docs"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=umls_mcp_server --cov-branch --cov-report=term-missing"
markers = [
    "integration: integration tests that call the real UMLS API (require UMLS_API_KEY)",
]
